refactor: cache -> strategy package 안으로 CacheStrategy interface 이동
