--- module-redis/src/main/java/com/github/msafriends/moduleredis/ModuleRedisApplication.java
@@ -1,13 +1,0 @@
-package com.github.msafriends.moduleredis;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class ModuleRedisApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(ModuleRedisApplication.class, args);
-    }
-
-}
