+++ module-redis/src/main/java/com/github/msafriends/moduleredis/cache/strategy/EvictionCacheStrategy.java
@@ -1,4 +1,4 @@
-package com.github.msafriends.moduleredis.cache;
+package com.github.msafriends.moduleredis.cache.strategy;
 
 public interface EvictionCacheStrategy<K, V> extends CacheStrategy<K, V>{
