+++ module-redis/src/main/java/com/github/msafriends/moduleredis/cache/strategy/LFUCacheStrategy.java
@@ -1,4 +1,4 @@
-package com.github.msafriends.moduleredis.cache;
+package com.github.msafriends.moduleredis.cache.strategy;
 
 /**
  * Interface representing Least Frequently Used (LFU) caching strategy.
