+++ module-redis/src/main/java/com/github/msafriends/moduleredis/cache/strategy/LRUCacheStrategy.java
@@ -1,4 +1,4 @@
-package com.github.msafriends.moduleredis.cache;
+package com.github.msafriends.moduleredis.cache.strategy;
 
 /**
  * Interface representing Least Recently Used (LRU) caching strategy.
