+++ module-redis/src/main/java/com/github/msafriends/moduleredis/cache/strategy/TimeToLiveCacheStrategy.java
@@ -1,4 +1,4 @@
-package com.github.msafriends.moduleredis.cache;
+package com.github.msafriends.moduleredis.cache.strategy;
 
 import java.time.Duration;
